# apps/YLunchApi.Main/wwwroot/swagger-original.yaml
...
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: rael.calitro@ynov.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://virtserver.swaggerhub.com/rael06/ylunch-api/1.0.0
    description: SwaggerHub API Auto Mocking
  - url: https://ylunch-api.rael-calitro.ovh
    description: Online server
  - url: http://localhost:5254
    description: Local server
security:
  - Bearer: [ ]
...
paths:
...
  /authentication/login:
    post:
      tags:
        - authentication
      summary: Login a user and returns access and refresh tokens as Json Web Token
      description: |
        - Authorization header: **no**
      operationId: login
      requestBody:
        required: true
        description: A JSON object that contains the user name and password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
            examples:
              restaurantAdmin:
                value:
                  email: admin@restaurant.com
                  password: Password1234.
              customer:
                value:
                  email: anne-marie.martin@ynov.com
                  password: Password1234.
      responses:
        '200':
          description: |
            Success
            - The access and refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenReadDto'
        '400':
          description: |
            Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                ValidationError:
                  value:
                    title: "One or more validation errors occurred."
                    statusCode: 400
                    errors:
                      Email: [ "The Email field is required." ]
                Other:
                  value:
                    title: "BadRequest"
                    statusCode: 400
                    errors:
                      reasons: [ "Something went wrong, please try again later." ]
        '401':
          description: |
            Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              example:
                title: "Unauthorized"
                statusCode: 401
                errors:
                  reasons: [ "Please login and use provided tokens." ]
  /authentication/refresh-tokens:
    post:
      tags:
        - authentication
      summary: Refresh user's tokens and returns access and refresh tokens as Json Web Token
      description: |
        - Authorization header: **no**
      operationId: refreshTokens
      requestBody:
        required: true
        description: A JSON object that contains the access and refresh tokens
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenUpdateDto'
      responses:
        '200':
          description: |
            Success
            - The new access and refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenReadDto'
        '400':
          description: |
            Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                ValidationError:
                  value:
                    title: "One or more validation errors occurred."
                    statusCode: 400
                    errors:
                      AccessToken: [ "The AccessToken field is required." ]
                Other:
                  value:
                    title: "BadRequest"
                    statusCode: 400
                    errors:
                      reasons: [ "Something went wrong, please try again later." ]
        '401':
          description: |
            Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              example:
                title: "Unauthorized"
                statusCode: 401
                errors:
                  reasons: [ "Invalid tokens, please login to generate new valid tokens." ]
  /authentication/current-user:
    get:
      tags:
        - authentication
      summary: Gets the current user based on the access token in headers
      description: |
        - Authorization header: **yes**
        - Granted roles: 
          - All
      operationId: getCurrentUser
      responses:
        '200':
          description: |
            Success
            - The current user Json object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserReadDto'
        '400':
          description: |
            Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              example:
                title: "BadRequest"
                statusCode: 400
                errors:
                  reasons: [ "Something went wrong, please try again later." ]
        '401':
          description: |
            Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              example:
                title: "Unauthorized"
                statusCode: 401
                errors:
                  reasons: [ "Please login and use provided tokens." ]
  /authentication/logout:
    get:
      tags:
        - authentication
      summary: Logs out the current user
      description: |
        - Authorization header: **yes**
        - Granted roles: 
          - All
      operationId: logout
      responses:
        '204':
          description: |
            NoContent
        '400':
          description: |
            Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              example:
                title: "BadRequest"
                statusCode: 400
                errors:
                  reasons: [ "Something went wrong, please try again later." ]
        '401':
          description: |
            Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              example:
                title: "Unauthorized"
                statusCode: 401
                errors:
                  reasons: [ "Please login and use provided tokens." ]
...
components:
  schemas:
...
    ErrorDto:
      properties:
        title:
          type: string
          description: The title.
        statusCode:
          type: number
          description: The status code.
        errors:
          properties:
            reasons:
              type: array
              items:
                type: string
                description: The reason.
    LoginRequestDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: anne-marie.martin@ynov.com
        password:
          type: string
          example: Password1234.
      required:
        - email
        - password
    TokenReadDto:
      type: object
      properties:
        accessToken:
          type: string
          example: |
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJJZCI6ImFlNTcxOGJkLWY5NzktNDQ1Ni04MmRlLTBjNzgzMzk4NzYzNCIsImp0aSI6IjBjZjc3NGYwLTUxYWQtNDM4Yi1hNTZlLTI1N2VlZGRhMjQ0MCIsIm5iZiI6MTY0MTEzNzMxMCwiZXhwIjoxNjQxMTU4OTEwLCJpYXQiOjE2NDExMzczMTB9.q1VaAodjovu79AyBNbrSxu4yZ4vOlDZQF2IyDvcsbJM
        refreshToken:
          type: string
          example: abcdefghij-a00000e0-aaa0-000a-0000-00e00aa00e00
    TokenUpdateDto:
      type: object
      properties:
        accessToken:
          type: string
          example: |
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJJZCI6ImFlNTcxOGJkLWY5NzktNDQ1Ni04MmRlLTBjNzgzMzk4NzYzNCIsImp0aSI6IjBjZjc3NGYwLTUxYWQtNDM4Yi1hNTZlLTI1N2VlZGRhMjQ0MCIsIm5iZiI6MTY0MTEzNzMxMCwiZXhwIjoxNjQxMTU4OTEwLCJpYXQiOjE2NDExMzczMTB9.q1VaAodjovu79AyBNbrSxu4yZ4vOlDZQF2IyDvcsbJM
        refreshToken:
          type: string
          example: abcdefghij-a00000e0-aaa0-000a-0000-00e00aa00e00
      required:
        - accessToken
        - refreshToken
...
    UserReadDto:
      type: object
      properties:
        id:
          type: string
          example: a00000e0-aaa0-000a-0000-00e00aa00e00
        email:
          type: string
          format: email
          example: user@domain.com
        firstname:
          type: string
          example: Anne-Marie
        lastname:
          type: string
          example: Martin-Jacques
        phoneNumber:
          type: string
          format: phone
          example: '0687654321'
        roles:
          type: array
          example:
            - RestaurantAdmin
          items:
            type: string
...
  securitySchemes:
    Bearer:
      type: http
      description: |
        JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'.
      scheme: bearer